+++ Patches/ImpactDetectorPatch.cs
@@ -32,9 +32,11 @@ namespace ForgottenDelivery.Patches
                     else
                         ForgottenDeliveryMod.log.LogInfo($"A regular package was destroyed, spawning a {items[index].name}.");
                     if (SemiFunc.IsMultiplayer())
-                        PhotonNetwork.InstantiateRoomObject("Items/" + items[index].prefab.name, __instance.transform.position, __instance.transform.rotation);
+                        PhotonNetwork.InstantiateRoomObject("Items/" + items[index].prefab.name, __instance.transform.GetChild(0).position, __instance.transform.GetChild(0).rotation);
                     else
-                        Object.Instantiate(items[index].prefab, __instance.transform.position, __instance.transform.rotation);
+                        Object.Instantiate(items[index].prefab, __instance.transform.GetChild(0).position, __instance.transform.GetChild(0).rotation);
+                    if (ConfigManager.keepItemsAfterLeaving.Value)
+                        StatsManager.instance.ItemPurchase(items[index].itemAssetName);
                 }
                 else if (bigPack)
                     ForgottenDeliveryMod.log.LogInfo("A big package was destroyed, but the current drop settings do not allow for an item to be dropped by this package.");