+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## ForgottenDelivery [1.0.2]
+- Added config setting to toggle keeping items dropped by packages
+- Attempt to fix issue of items spawning in walls by adjusting the spawn position and rotation
+
 ## ForgottenDelivery [1.0.1]
 - Added blacklist config options to allow for specific items to be banned from dropping
 - Fixed desync issues caused by the photon view having the wrong components